+++ contracts/interfaces/IHouseOfCoin.sol
@@ -27,11 +27,7 @@ interface IHouseOfCoin {
      * @param amount To mint.
      * Emits a {CoinMinted} event.
      */
-    function mintCoin(
-        address reserveAsset,
-        address houseOfReserve,
-        uint amount
-    ) external;
+    function mintCoin(address reserveAsset, address houseOfReserve, uint256 amount) external;
 
     /**
      * @notice  Function to payback ERC20 'backedAsset' of this HouseOfCoin.
@@ -40,27 +36,21 @@ interface IHouseOfCoin {
      * @param amount To payback.
      * Emits a {CoinPayback} event.
      */
-    function paybackCoin(uint _backedTokenID, uint amount) external;
+    function paybackCoin(uint256 _backedTokenID, uint256 amount) external;
 
     /**
      * @notice  External function that returns the amount of backed asset coins user can mint with unused reserve asset.
      * @param user to check minting power.
      * @param reserveAsset Address of reserve asset.
      */
-    function checkMintingPower(address user, address reserveAsset)
-        external
-        view
-        returns (uint);
+    function checkMintingPower(address user, address reserveAsset) external view returns (uint256);
 
     /**
      * @notice  Function to get the health ratio of user.
      * @param user address.
      * @param houseOfReserve address in where user has collateral backing debt.
      */
-    function computeUserHealthRatio(address user, address houseOfReserve)
-        external
-        view
-        returns (uint256);
+    function computeUserHealthRatio(address user, address houseOfReserve) external view returns (uint256);
 
     /**
      * @dev Returns the _liqParams as a struct
