+++ contracts/interfaces/IWETH.sol
@@ -4,7 +4,7 @@ pragma solidity 0.8.17;
 interface IWETH {
     function deposit() external payable;
 
-    function transfer(address to, uint value) external returns (bool);
+    function transfer(address to, uint256 value) external returns (bool);
 
-    function withdraw(uint) external;
+    function withdraw(uint256) external;
 }
